Updated db schema
Changed primary key to a serialized integer. Decided this would be safer than relying upon the imdb_id (as this, unlikely as it is, may change). Created an imdb_id column, which I then indexed for faster lookup.

Added additional columns for all the info I would like to display about a film. Once a user rates the film and it is added to the database, the next time the film details are requested, an api call will not need to be made