Route to get film detail completed
A request can be made to /films/:imdbID to get details about the film. If the film is already in the database, that record will be returned. Otherwise, an api call is made, and then the response is formatted before being added to the database. The record in the database is then returned as the response object